Chroma#536: Minor test refactoring
- changed chromaclient and app.ts path in tests
