+++ clients/js/test/collection.client.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "@jest/globals";
+import { expect, test, beforeEach } from "@jest/globals";
 import chroma from "./initClient";
 
 beforeEach(async () => {
